requirements.txt
requests
python-dotenv
pycoingecko
matplotlib
pillow
